test(stdlib): split DistanceTo and GradientCast tests
