+++ stdlib/src/commonTest/kotlin/it/unibo/collektive/stdlib/test/GradientCastTest.kt
@@ -12,83 +12,20 @@ import it.unibo.collektive.aggregate.api.Aggregate
 import it.unibo.collektive.aggregate.api.neighboring
 import it.unibo.collektive.stdlib.spreading.GradientPath
 import it.unibo.collektive.stdlib.spreading.bellmanFordGradientCast
-import it.unibo.collektive.stdlib.spreading.distanceTo
 import it.unibo.collektive.stdlib.spreading.gradientCast
 import it.unibo.collektive.stdlib.util.euclideanDistance3D
 import it.unibo.collektive.testing.Environment
 import it.unibo.collektive.testing.EnvironmentWithMeshNetwork
 import it.unibo.collektive.testing.Position
 import it.unibo.collektive.testing.SerializingMailbox
-import it.unibo.collektive.testing.mooreGrid
-import kotlin.math.abs
 import kotlin.math.nextDown
 import kotlin.math.nextUp
-import kotlin.math.sqrt
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertFalse
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
-class SpreadingTest {
-    private fun Environment<Double>.gradientIsStable(isMoore: Boolean, size: Int): Boolean =
-        status().all { (id, value) ->
-            val x = id % size
-            val y = id / size
-            val steps = x + y
-            val expected =
-                when {
-                    isMoore -> {
-                        val manhattanSteps = abs(x - y)
-                        check((steps - manhattanSteps) % 2 == 0)
-                        val diagonalSteps = (steps - manhattanSteps) / 2
-                        sqrt(2.0) * diagonalSteps + manhattanSteps
-                    }
-
-                    else -> steps
-                }
-            abs(value - expected.toFloat()) < 1e-6
-        }
-
-    private fun mooreGridWithGradient(size: Int) =
-        mooreGrid<Double>(size, size, { _, _ -> Double.NaN }) { environment ->
-            val localPosition = environment.positionOf(localId)
-            distanceTo(localId == 0) { euclideanDistance3D(localPosition.toTriple()) }
-        }.apply {
-            assertEquals(size * size, nodes.size)
-            val initial = status().values.distinct()
-            assertEquals(1, initial.size)
-            check(initial.first().isNaN()) {
-                "Initial status is not NaN, but it is $initial (${initial::class.simpleName})"
-            }
-        }
-
-    @Test
-    fun `distanceTo in the luckiest case stabilizes in one cycle`() {
-        val environment: Environment<Double> = mooreGridWithGradient(10)
-        environment.cycleInOrder()
-        assertTrue(environment.gradientIsStable(isMoore = true, 10))
-    }
-
-    @Test
-    fun `distanceTo requires at most the longest path length cycles to stabilize`() {
-        val size = 5
-        val environment: Environment<Double> = mooreGridWithGradient(size)
-        environment.cycleInReverseOrder()
-        val firstRound = environment.status()
-        assertEquals(0.0, firstRound[0])
-        firstRound.forEach { (id, value) ->
-            when (id) {
-                0 -> assertEquals(0.0, value)
-                else -> assertEquals(Double.POSITIVE_INFINITY, value)
-            }
-        }
-        repeat(size - 1) {
-            assertFalse(environment.gradientIsStable(isMoore = true, size))
-            environment.cycleInReverseOrder()
-        }
-        assertTrue(environment.gradientIsStable(isMoore = true, size))
-    }
+class GradientCastTest {
 
     @Test
     fun `GradientPath is serializable`() {
